feat(tup-cms): tup-387, list plugin
